+++ packages/lint-suite/src/lib/javascript.ts
@@ -1,3 +1,8 @@
-import nx from '@nx/eslint-plugin';
+import { configs } from '@nx/eslint-plugin';
+import { type ConfigArray, config } from 'typescript-eslint';
 
-export const javascript = nx.configs['flat/javascript'];
+export const javascript: ConfigArray = config({
+  files: ['**/*.js', '**/*.jsx', '**/*.cjs', '**/*.mjs'],
+  extends: [...configs['flat/javascript']],
+  rules: {}
+});